Import setuptools before distutils
Setuptools wants to be imported before distutils, and prints a warning if it
is not. Starting with a recent version (>57.2, <=61.2), not following this
rule causes catkin's own monkey patching to fail. This commit reverses the
import order to fix the warning and patching issue.

Comment by v4hn: While this does not actually trigger failures on most systems
setuptools clearly documents that setuptools should be included *before* distutils.

Co-authored by v4hn <me@v4hn.de>